requirements.txt
beautifulsoup4==4.12.3
brotlipy==0.7.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
EbookLib==0.18
idna==3.10
lxml==5.3.0
pillow==10.4.0
pycparser==2.22
requests==2.32.3
six==1.16.0
soupsieve==2.6
urllib3==2.2.3
